+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Modified alpha level 1 to focus on jumping.
 - Player now moves automatically when the level starts.
 - Diving into an enemy now sets the player's state to airborne, allowing them to use any other move.
+- No starting point object in the tilemap will no longer result in a crash.
 
 ### Added